Safari-609.1.11.1/LayoutTests/css3/flexbox/canvas-dynamic-change.html

<!DOCTYPE html>
<link href="resources/flexbox.css" rel="stylesheet">
<style>
.red {
  height: 400px;
  background: red;
}
canvas {
  background: green;
}
</style>
<script src="../../resources/testharness.js"></script>
<script src="../../resources/testharnessreport.js"></script>
<script src="../../resources/check-layout-th.js"></script>
<div id=log></div>

<div class="red">
  <div class="flexbox column">
    <canvas id="canvas" data-expected-height="400"></canvas>
  </div>
</div>

<script>
var canvas = document.getElementById('canvas');

requestAnimationFrame(function() {
  canvas.height = 400;
  checkLayout('.flexbox');
});

</script>
